refactor: assessment types
