+++ app/data/assessmentTypes.ts
@@ -1,0 +1,13 @@
+const ASSESSMENT_TYPES = {
+  projectManagement: {
+    type: 'Project Management assessment',
+    slug: 'project-management',
+  },
+  permitting: { type: 'Permitting assessment', slug: 'permitting' },
+  technical: { type: 'Technical assessment', slug: 'technical' },
+  gis: { type: 'GIS assessment', slug: 'gis' },
+  financialRisk: { type: 'Financial Risk assessment', slug: 'financial-risk' },
+  screening: { type: 'Eligibility Screening', slug: 'screening' },
+};
+
+export default ASSESSMENT_TYPES;
